Moved writeJar from Session to REPL - not sure it belonged in Session
